Remote: extract the chunked-transfer receiver out of the command router
Moves the Debrify-channel chunked transfer machinery into
`lib/services/remote_control/remote_chunk_transfer.dart` as
`RemoteChunkTransferReceiver`, a real object that owns its own buffers:
`_handleDebrifyChannelStart` -> `handleStart`, `_handleDebrifyChannelChunk`
-> `handleChunk`, plus `_armChunkTimeout`, `_takeReassembledPayload`,
`_completeEncryptedBlob`, `_reportChunkFailure` and `_ChunkBuffer`. The
per-source drop tally behind the "update your phone app" notice
(`_noteUnauthenticatedDrop`, `_StaleRemoteSource`) moves alongside it into
`remote_stale_source_notice.dart` as `RemoteStaleSourceNotice`.

Both units take everything they need through constructor callbacks — the
snackbar sink, command dispatch, the peer-key derivation and the two
best-effort result reporters. Neither file references a router private, and
neither imports Flutter widgets, `lib/widgets` or `lib/screens`; the router's
own `_RouterBusyDialog` violation is untouched.

Bodies move verbatim apart from those renames; `dart format` reflowed the
`handleStart` signature onto one line because the shorter name fits. The
16 MB payload ceiling becomes `kMaxRemoteTransferPayloadBytes` in the new
file and the router's `_maxProfileRemotePayloadBytes` aliases it, so the
staged-profile buffer keeps measuring the same budget.

Quirks preserved: the plain (v1) single long stall deadline and its
receiver-side notice, the v2 gap-repair rounds and their diagnostics, the
duplicate-chunk and repeated-start no-ops, silent refusal of a transfer
naming a transport command, and the stale-source burst threshold/cooldown.

Pinned by `test/remote_chunk_transfer_origin_pin_test.dart` (11 tests, added
in the previous commit against the origin path); it passes unchanged after
the move.

Host: 5016 -> 4300 lines. No forwarders.
